+++ src/me/devtec/servercontrolreloaded/utils/playtime/PlayTimeUtils.java
@@ -21,8 +21,8 @@ import me.devtec.theapi.utils.theapiutils.LoaderClass;
 
 public class PlayTimeUtils {
 	
-	public static HashMap<UUID, Long> playtop = new HashMap<>(); //Player UUID || PlayTime
-	public static RankingAPI<UUID, Long> ranks = new RankingAPI<>(playtop);
+	public static HashMap<UUID, Integer> playtop = new HashMap<>(); //Player UUID || PlayTime
+	public static RankingAPI<UUID, Integer> ranks = new RankingAPI<>(playtop);
 	
 	static Statistic st;
 	
@@ -49,7 +49,7 @@ public class PlayTimeUtils {
 		else
 			return API.getSPlayer(s).getPlayTime("PlayTime");
 	}
-	public static long playtime(String player) {
+	public static int playtime(String player) {
 		OfflinePlayer s = Bukkit.getOfflinePlayer(player);
 		if(s==null)return -1;
 		if( !Loader.config.getBoolean("Options.PlayTime.UseCustomPlayTime"))
